Fixing the definition of OldShearMethod to be int, consistent with its usage
--HG--
branch : week-of-code

+++ src/enzo/global_data.h
@@ -576,7 +576,7 @@ EXTERN float MinimumShearForRefinement;
 /* For CellFlaggingMethod = 9,   
    Whether to use the old method for calculating shear refinement.    */
 
-EXTERN float OldShearMethod;
+EXTERN int OldShearMethod;
 
 /* For CellFlaggingMethod = 11,
    The number of cells by which the Resistive length abs(B)/abs(curl(B))
